Create HTML::MasonX::ApacheLikePlackHandler (again)
Get rid of a pointless eval() that was only here in the original
HTML::Mason::ApacheHandler because it needed both Apache 1 and Apache 2
method compatibility. By not making it an eval I'll get meaningful
stacktraces for this, not just an "(eval)" stackframe.

Also just remove the ":method" signature. Since we're not under Apache
nothing's going to care about it.

Should have just been done in the original creation of this module in
91887cb, but I missed it at the time.